Chameleon: Abstract the screen comparison logic to the ScreenComparer class
This change abstracts the screen comparison logic to a new class, i.e.
ScreenComparer. It is created via the ScreenUtilityFactory, depending on
the connected interface is VGA or not. Proper configurations, like
pixel_diff_margin and wrong_pixels_margin, are preset.

It also abstracts the mirror comparison logic which checks all screens
on CrOS identical or not.

BUG=chromium:429931
TEST=Tested on Peppy by running Chameleon tests. Tests passed.
TEST=Tested on Parrot by running Chameleon tests. Tests passed.

Change-Id: I3c43269a517e3afd1e78f6a314fc33d9ddda9a6f
Reviewed-on: https://chromium-review.googlesource.com/229669
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>