Fix unittest problem uncovered with GCC 4.9
When building 64-bit boards using GCC 4.9, we uncovered a unittest
problem where the unittest was incorrectly depending on python
dictionary entries always being in the same order (dictionaries are
supposed to be unordered).

This CL fixes the bug.

BUG=chromium:430296
TEST=Tested this fix both with GCC 4.9 build, when test failed without the
fix; and with GCC 4.8 build, to verify GCC 4.8 build continues to work
with this fix.

Change-Id: I083a2c9eb6b2ba21179d3227378805c3591c0c33
Reviewed-on: https://chromium-review.googlesource.com/229752
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>