[autotest] Do not fail cleanup if it fails to restore RPM but battery level is above 50%.
If the code fails set RPM to back on in cleanup, it will check if battery level
is high enough (>50%) to run tests, if so, no exception will be raised, so the
next reset job can try to restore RPM again.

The code als checks if AC is actually connected and charging, if so, remove
`outlet_changed` host attribute, so no more attempt to restore RPM will be
wasted.

BUG=chromium:433398
TEST=local setup
force power_on to raise exception and make sure cleanup works.

Change-Id: Ic546dbcb175115f94605ff4adc375651b630b6d2
Reviewed-on: https://chromium-review.googlesource.com/229858
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>