unittest_suite: Add exception for mbim_compliance modules.
There are two modules in client/cros/cellular/mbim_compliance that are named
'*_test.py' but aren't really tests. These modules are in the process of going
away. Add temporary exception to not treat them as tests to ease the process of
their removal.

BUG=chromium:432621
TEST=`./utils/unittest_suite.py -r client/cros/cellular/mbim_compliance` passes.

Change-Id: Ib25758a32a64f5845d253a677d32706f5f3424bd
Reviewed-on: https://chromium-review.googlesource.com/229960
Reviewed-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>