autotest: move wake-on-WiFi features to new shill API
Wake-on-packet is now configured on a per-device basis, so we
should move it out of Manager. Additionally, the service order
context manager is no longer necessary. Instead, there is a
manager for the wake-on-WiFi features to prevent us from
modifying global state across multiple tests.

This also updates the WoWLAN test to use the new API.

BUG=chromium:426222
TEST=test_that with panther AP

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Change-Id: I4c6ba1b743a50f023a68bfa499b65e872a6d5811
Reviewed-on: https://chromium-review.googlesource.com/230261
Reviewed-by: Christopher Wiley <wiley@chromium.org>