[autotest] Refactor HostJobHistory creation in dut_status
Previously, a HostJobHistory object was created by calling the
constructor with a hostname.  This changes the constructor to expect
an instance of an AFE Host model object.  Creating HostJobHistory
instances by hostname is now done with a factory method on the
class.

A side effect of the change is to improve error reporting when the
command line includes an unknown host.

This change also includes completely gratuitous white space changes
in the module's docstring.

BUG=None
TEST=run the command with valid and invalid host names

Change-Id: I514123d41a45fb42e12369ce15d230247138ec69
Reviewed-on: https://chromium-review.googlesource.com/230548
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>