Chameleon: Don't test VGA EDIDs on a HDMI interface
The HDMI receiver only accepts DVI in addition, but not VGA. Testing VGA
EDID on a HDMI interface does not make sense. We will test them either on
a native VGA interface or via a HDMI to VGA adapter.

Also add the VGA configuration in the test.

BUG=chromium:435356,chromium:429931
TEST=Ran the display_EdidStress on a DUT with HDMI passed.

Change-Id: I905f08cd66524f8a1f8d86c50447c73a506e6fc1
Reviewed-on: https://chromium-review.googlesource.com/231158
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>