[autotest] Drop suite_min_duts keyval when creating suite job
This is part III of making host scheduler support a suite_min_duts
requirement per suite.

With this CL, the suite_min_duts is piped from run_suite
to the rpc method create_suite_job and dropped as a job keyval.

DEPLOY=apache
TEST=Run run_suite, confirm 'suite_min_duts' is dropped
as a job keyval of the suite in the db.
BUG=chromium:432652

Change-Id: I3193aba8c0938c3e8d7b3854c19b42c3f14e2829
Reviewed-on: https://chromium-review.googlesource.com/232570
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>