[autotest]: Fix revision_control_unittests for hostname-less Systems.
The revision_control_unittests create a local checkout and tests
manipulating it. However it does not set the user name or email,
therefore in most cases the commit test automatically generates
the author info. This however fails on systems that don't set
the hostname.

Therefore the unittests now set up a dummy user name and email for
the test checkout.

BUG=chromium:436739
TEST=Autotest unittests now pass on mobbuild.

Change-Id: I6b0f32de70f24e6e64c934519afcf9844a19a5ac
Reviewed-on: https://chromium-review.googlesource.com/232140
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>