video_VideoDecodeMemoryUsage - Log the memory usage incremental rates.
In this test, we measure the total memory usage while video playing
back. However, if there is any change in chromium that affects the
memory usage. It will trigger the alarms of this test.

To reduce the number of these false alarms, we change to measure the
incremental rate instead. However, the test may still fail if we
leak memory in other modules like opening and closing tabs.

Also log the measured numbers so that we can do more analysis.

BUG=chromium:306347
TEST=Run the test on peach_pit

Change-Id: Ia3260a63676ef54167142f21ccb96cf0fcb53482
Reviewed-on: https://chromium-review.googlesource.com/232073
Reviewed-by: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>