video_VideoDecodeMemoryUsage - Disable the DEBUG message in telemetry
In the autotest, the autoserv logs its stdout to a file under
"/tmp/autoserv-XXXXX/". However, the partition /tmp is actually of
type tmpfs and takes memory space when files are stored on it.

The entry "Shmem" in "/proc/meminfo" shows the usage. Therefore, the
test thought there are memory leak. To fix this, we disable the debug
information in telemetry module to reduce the number of logging
messages.

BUG=chromium:331992
TEST=Run the test on peach pit

Change-Id: I01775b4feccf2b72f5f4654f6c418d3b1a7d3da9
Reviewed-on: https://chromium-review.googlesource.com/232240
Reviewed-by: Kuang-che Wu <kcwu@chromium.org>
Reviewed-by: Owen Lin <owenlin@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>