bluetooth_socket: fix wrong controller index error
When making a command to the kernel, rather than the controller, e.g.
"version" or "read supported commands" we supply an index of 0xffff
and expect that in the response just as we expect 0x0 for the first
controller.

However the kernel may be reporting events for the controller itself
inbetween the command and the response, which will trip this error
check.

Move the error check inside the check for the command complete/status
events to fix that.

BUG=none
TEST=add a version check after a reset() call

Change-Id: I23eebceafbff8ef04dae29f79e475b13d0525c17
Reviewed-on: https://chromium-review.googlesource.com/232440
Reviewed-by: Arman Uguray <armansito@chromium.org>
Reviewed-by: Michael Janssen <jamuraa@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>