bluetooth_tester: don't turn HS off
Intel-based testers have the HS flag on, and cannot be turned off,
while Atheros-based testers have it turned off by default.

Deal with this difference by turning HS on if the profile demands it,
but not turning it off if the profile doesn't.

BUG=chrome-os-partner:33379
TEST=use a panther tester

Change-Id: Ic1b27fddbd66e71479ddc378119db3db929a1d4f
Reviewed-on: https://chromium-review.googlesource.com/232523
Tested-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Michael Janssen <jamuraa@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>