Check for disconnects in SimpleConnect test
Assure that the client does not disconnect during the duration
of the pings in all WiFi_SimpleConnect tests.  In addition to
covering some failure modes in the checkTruncatedBeacon test,
this is also a resonable bar to hold all other SimpleConnect
tests to.

BUG=chrome-os-partner:33662
TEST=Run this test, on both ath9k and WP2 based cells.  Also
run check11b and ensure it continues to pass on all boards.

Change-Id: Ibff56a3056c7d172294fab422fc6274abaaca37c
Reviewed-on: https://chromium-review.googlesource.com/233221
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>