bluetooth_tester: Fix Intel WP2 as Tester
It seems that Intel WP2 in general does not like settings being changed
while powered on, so reorganize the setup to set the basic settings first,
then power on the adapter, then finally set the fast-connectable setting
(which does require adapter power).

In addition, force power off the adapter even if already powered off; this
somehow fixes the AP-Mode variant of WP2 to accept settings.

With these changes, the HS mode can actually be toggled on/off, so revert
the previous patch that kept it on if already on.

BUG=chrome-os-partner:33379
TEST=test_that suite:bluetooth_sanity

Change-Id: Idacb6e1b36eabeaf00ebea33c33e0cb02364c7da
Reviewed-on: https://chromium-review.googlesource.com/232776
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Michael Janssen <jamuraa@chromium.org>
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>