video_VDAPerf - Warm up the rendering to have better performance.
In the vda_unittest, it would be slow to draw the first few frames.
The reason is we need some warming up after initialize the GL
environment. In Chrome, before playing the video, the graphics has
been used for a long time. It is reasonable to do the warm up in the
test. Without the warming up, about 5% frames would be dropped
because it needs to catch up the initial slowness.

BUG=chromium:411123
TEST=Run the test on peach pit

Change-Id: I580d34e7641d04dd7469469bf30661857b8a9d66
Reviewed-on: https://chromium-review.googlesource.com/233115
Reviewed-by: Kuang-che Wu <kcwu@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>