mbim_compliance: Add test CM_15 specified in the compliance suite.
1. Test checks whether the device correctly fragments the control
response
messages when the host sets the max_control_transfer parameter to a
low value in the open sequence.
2. Adding an optional parameter to open sequence invocation to set the
max_control_transfer_size parameter. If not explicitly set by the test,
it defaults to the wMaxControlMessage parameter specified by the device
in the functional descriptor as before.
3. Modifying all existing tests, sequences to use the configure
max_control_transfer size setting for all control transfer operations
rather than assuming it to be set to wMaxControlMessage value specified
in the device functional descriptor.

BUG=chromium:432621
TEST=test_that -b ${board} ${host} suite:network_mbim_compliance

Change-Id: Ic32e551f3dd02f34ad05a2b199405b3a41a6a1b7
Reviewed-on: https://chromium-review.googlesource.com/233231
Reviewed-by: Roshan Pius <rpius@chromium.org>
Commit-Queue: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>