Bluetooth: allow tester IP to be specified in --args
Normally the tester hostname is derived from the DUT hostname by
appending '-router' to the first part. This requires manual /etc/hosts
entries on your workstation when dealing with machines by IP address.

Allowing the IP (or hostname) of the tester to be specified on the
command-line with '--args tester=IP' makes this much easier to handle.

BUG=none
TEST=test_that $DUT_IP bluetooth_Sanity_Discovery --args tester=$TESTER_IP

Change-Id: Ic113cb73a2de95e60b75931d95f2b1ea7ba4a2cf
Reviewed-on: https://chromium-review.googlesource.com/233492
Reviewed-by: Michael Janssen <jamuraa@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>