Chameleon: Simplify and refine the logic of applying an EDID
It is hard to apply an EDID in a correct way. To emulate a real
scenario of switching monitor, it should:
 - unplug the monitor,
 - apply the EDID,
 - plug the monitor, and
 - wait the source signal stable.
And also restores to the original EDID after a test.

To simplify these steps, a test can apply an EDID like the following:
    with chameleon_port.use_edid_file(filename):
        do_some_test_on(chameleon_port)

BUG=chromium:429862
TEST=Ran the affected tests passed on nyan_big.

Change-Id: Id22ff52341ecdc138879b5c8178d808fe15ebadd
Reviewed-on: https://chromium-review.googlesource.com/233760
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>