security_OpenFDs: fix on ASan.
We used to need to whitelist /proc when running on ASan,
but it seems to not be needed anymore. Keep the whitelist entry for now,
but make the test not fail when there's unmatched whitelist entries.

Extract the ASan check to its own file.

Also clean up some pre-submit errors.

BUG=chromium:446885
TEST=Passes on expresso w/ASan.

Change-Id: I250cad966bc83cda4480322e6cc746a9ecf9143c
Reviewed-on: https://chromium-review.googlesource.com/241119
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>