Detect if TPM is under a dictionary attack.
A lot of time is being spent triaging failures due to the fact that the
TPM is in defend mode.  Since possible error messages are "p11_replay",
"Command returned non-zero exit status" or others, update the test to
speed up triage.

If the TPM has a dictionary attack count greater than 0, throw TestError
with a message that says so.  That way when the bug is automatically
filed that message will be in the anchor.

BUG=chromium:440241
TEST=Manual; Ran it

Change-Id: I56b98b43b66140079ac611a46563cf0cf0556c45
Reviewed-on: https://chromium-review.googlesource.com/234041
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Tien Chang <tienchang@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>