[autotest] Redirect log links to shards.
This change checks the shadow_config for a list of shards and
looks for logs on the shards the same way we look for logs on
drones. This is convenient in that any client of these logs
can use a generic url like:
  http://cautotest/tko/retrieve_logs.cgi?job=/results/<job_id>-<user>
and the retrieve_logs script on cautotest will handle translating
shard hostnames. This is a working solution, but it isn't ideal
because polling all shards for log urls won't scale. Unfortunately
it is how we've handled drones till now, so we can evolve this
solution as we roll out shards.

TEST=Checked log links on puppylab.
BUG=chromium:438858,chromium:438857

Change-Id: Ie58b88ea41c33c3148098e807ddd6ab7f8f4414f
Reviewed-on: https://chromium-review.googlesource.com/234291
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>
Trybot-Ready: Prashanth B <beeps@chromium.org>