deploy_production_local: Correctly accept prod branch as clean.
We were declaring the server cautotest as dirty because it has a local
'prod' branch. We've decided to keep this branch for now, so it should be
treated as clean.

Also redo some command output to make it more understandable, especially
when --dryrun is specified.

BUG=chromium:425636
TEST=Unittests + dryrun testing.

Change-Id: I2828d79f7f5bd56140103c99379d7f26e5be885a
Reviewed-on: https://chromium-review.googlesource.com/234820
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>