[autotest] Delay retrieving IP address of a host.
The error raised in initializer will lead to CrosHost failed to be initialized.
We need to delay this action only when it's requested. so CrosHost can be
initialized and continue working for cases like repairing, in which case the
dut can be down and don't have an IP address if DHCP server does not have
static IP for it.

BUG=chromium:441511
TEST=local test.

Change-Id: I3e83c4a9ddfd00add4b373f5a2c9339d77f61aa6
Reviewed-on: https://chromium-review.googlesource.com/234851
Trybot-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>