Adding a control file for stress lab testing.
This control file is needed, because it provides the stress lab to run
more fine tuned tests than in existing control files.

The video based control file previously was added to the
stress_experimental suite, but the video tests are having issues, so
adding this control file which runs the bare suspend resume is valuable
till the other issues are sorted out.

It also allows for us to run custom durations in the stress lab.

BUG=None
TEST=None
DEPLOY=test_importer

Change-Id: I5678c303c51cb9309b7fc0270ee990126b045045
Reviewed-on: https://chromium-review.googlesource.com/234891
Reviewed-by: harpreet Grewal <harpreet@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>