bluetooth_tester: fix class of device setup
Move the setting of Class of Device and Local name to before powering
up the adapter, so they are configured automatically as a part of that.

This solves a race condition where after being setup once, a tester
fails to be initialized due to a 'Busy' condition.

BUG=chromium:442016
TEST=run a test twice in a row without rebooting tester

Change-Id: I43acdd9deb4b92e9d2d4bee71a91d4f72f43d497
Reviewed-on: https://chromium-review.googlesource.com/235877
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>