full_release_test.py: Handle delta payload file names on non-release builds
Non-release builds have a date fragment in the version file name. Handle
it as well.

This change was previously reviewed on https://chromium-review.googlesource.com/236675 ,
landed and reverted. This patch includes a fix for the case where the
source and target version are the same (npo) in release builds.

BUG=chromium:425404
BUG=chromium:390828
TEST=New unit test
TEST=Manual test (see below)

Manual test:
* Regression test on release builds:
$ site_utils/autoupdate/full_release_test.py --npo --nmo --dump --dump_dir /tmp/au_test --archive_url gs://chromeos-image-archive/swanky-release/R41-6630.0.0 6630.0.0 swanky '--log=debug' --dry_run
This command generates a control file in /tmp/au_test with no
differences in the uris when comparing it with the one stored in
gs://chromeos-image-archive/swanky-release/R41-6630.0.0/au_control.tar.bz2
which was green. This fixes the regression introduced in CL:236675
(already reverted) where the source_image_uri was missing the full
payload filename.

* Trybot test:
Ran a lumpy-release trybot with --hwtest enabled, that uploaded the
build artifacts to
gs://chromeos-image-archive/trybot-lumpy-release/R43-6831.0.0-b2425
where the version will include the '-a1' suffix in the payloads.
Ran the same command using the trybot urls and checked the generated
source_image_uri variable in the control file has the right full payload
(found in the same artifacts directory):
$ site_utils/autoupdate/full_release_test.py --npo --nmo --dump --dump_dir /tmp/au_test --archive_url gs://chromeos-image-archive/trybot-lumpy-release/R43-6831.0.0-b2425 6831.0.0-a1 lumpy '--log=debug' --dry_run

Change-Id: I2ed839f16883130b5601701e99ebf76fad589312
Reviewed-on: https://chromium-review.googlesource.com/254860
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>