utils: Add dark_resume_utils for WoWLAN tests
For tests like wake-on-packet and wake-on-disconnect, we are
going to need to wake up in dark resume. However, it is possible
that USB devices will still be suspended in dark resume, and the
autotest server won't be able to communicate with the DUT when this
is the case.

So, we need to make sure we enable dark resume, and also pull the
USB ports out of the dark_resume_devices file so they will be
active in dark resume.

BUG=chromium:426222
TEST=using network_WiFi_WoWLAN, network_WiFi_WakeOnDisconnect

Change-Id: I2f7141d6c51b773be13216f8b85e88bbba0df13c
Reviewed-on: https://chromium-review.googlesource.com/236473
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Trybot-Ready: Eric Caruso <ejcaruso@chromium.org>