Update the stable repair build to R41-6592.0.0.
This is to pick up a working build for zako_freon.  Freon builds
earlier than 6592 had a bug that caused failures on any chromebox
in the lab.

Build 6592 was green for every board except stumpy_moblab.  It's
possible that we will need to adjust the stable build for the sake
of stumpy_moblab in the near future; this change risks the break
in moblab in order to get progress on freon.

BUG=chromium:436587
TEST=review wmatrix; install the zako_freon build on two DUTs

Change-Id: I351877d00ad017d72f11d0618ecba5a51c0d452b
Reviewed-on: https://chromium-review.googlesource.com/236616
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>