deploy_production_local.py: Remove repo terminal formatting.
The repo command sometimes inserts terminal formatting code. This CL strips
it out in the cases we care about via regex.

It was also possible to solve this by changing the repo config in
/usr/local/.repo/manifests/.git/config, but that seemed a bit harder
to maintain/distribute for now.

Also, adjust our final report to list the current git TOT versions if
we didn't do an update. This was done to help with testing this change,
but it's reasonable to have anyway.

BUG=chromium:443390
TEST=Local + Unitests + Manual on chromeos-autotest.cbf

Change-Id: I84cde1c9941c1b1a100e29b176587380b5994035
Reviewed-on: https://chromium-review.googlesource.com/236703
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>