Chameleon: Force to unplug after the test iteration
Most of the tests call iterate_all_ports of ChameleonVideoInputFinder
as its main loop logic. It used to restore the original plug state
after the iteration. However, some unknown reason leaves a display
port plugged. And it lasts for a long time and interferes with other
tests. It is better to force to unplug the port after the iteration,
making sure it not interfere others.

BUG=chromium:441882
TEST=Ran a Chameleon test and checked the plug state at the end.

Change-Id: I2c1470995ba4dc8684bddc162caf75132ab174b2
Reviewed-on: https://chromium-review.googlesource.com/236945
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>