autotest: Fix attenuator address inference for desktop testing
crbug.com/236657 broke WiFi testing outside the lab because test logic
throws an error when it sees that it cannot infer an attenuator DNS
name.  Add back logic to tolerate failures for attenuator DNS name
inference only.

BUG=None
TEST=manually launched tests against desktop setups decide that
that attenuators are not present and that it is ok.

Change-Id: I2cc07f9ccc45ba94d00846e167c04ff49f60dd0a
Reviewed-on: https://chromium-review.googlesource.com/237311
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>