[autotest] Allow shards to abort their tests and sync back status.
Currently if a job is running on a shard and we abort its parent,
eg: via abort_suite.py, the children jobs on the shard *may not*
abort. They will abort if the shard does a heartbeat inbetween
when abort_suite sets the aborted bit on the hqes and the master
scheduler sets the completed bit (in _find_aborting).

This change makes it so the master scheduler ignores all jobs
that are running on shards. This way whenever the shard does a
heartbeat it will learn about the aborted bit set by abort_suite,
kill the jobs, and sync back the aborted status to the master.

BUG=chromium:444754
TEST=Aborted jobs running on shard via their parent suites
     on the master.
DEPLOY=scheduler,shard

Change-Id: I00784138e9080988651b463410668221cf6f0267
Reviewed-on: https://chromium-review.googlesource.com/237325
Trybot-Ready: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>
Trybot-Ready: Dan Shi <dshi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>