Fix graphics_utils.get_external_connector_name()
Originally the method may return the name of a connector that is still
in the transient state of being connected, which may cause errors in
subsequent operations; e.g., get_external_resolution() returning (0, 0).

This change fixes it by making sure that the returned connector is in
final connected state.

BUG=chromium:441468
TEST=Ran display_HotPlugNoisy.extended on nyan_big and peach_pit and
saw correct calibration image shown on the screen instead of a zero-sized
one which would result in a blank screen.

Change-Id: Ia77d02a1e5d2c23e0636ba253828a0e562960ef2
Reviewed-on: https://chromium-review.googlesource.com/237921
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>