Add recovery step to display_HotPlugNoisy
The current test is strict in that it fails if DUT behaves incorrectly
after applying the noise. Now consider what a normal user would do when
they encounter the situation: they would replug the cable and see if it
recovers!

This change emulates this behavior and makes the test pass if DUT does
recover after the cable is replugged.

BUG=chromium:441468
TEST=Ran display_HotPlugNoisy.extended/mirrored 20 times on nyan_big and
nyan_blaze and saw all of them passed with the new replug recovery appearing
almost in every run. Also ran the test on peach_pit without problem.

Change-Id: Iaee9f6739f909b5e9c55fec85d4bdfa92a57ca37
Reviewed-on: https://chromium-review.googlesource.com/237970
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>