graphite: Factor time_utils out of the es_utils module.
We're planning to move the graphite support code into chromite so it can be
used by autotest and by other infrastructure bits. To do that, we have to
remove dependencies on other bits of autotest, specifically time_utils and the
global configuration.

The time_utils module is used in essentially one spot where it converts a
function parameter into a particular format. To remove that use, the conversion
is pushed out to the callers. That's a bit clunky and cumbersome, but once the
code has been moved we can add a little wrapper function which will centralize
that again. I don't want to add the wrapper now because I'd want to put it
where the other code is currently.

BUG=chromium:446291
TEST=Ran a trybot for butterfly-paladin with --hwtest.

Change-Id: Iaf2ee2e664ff8040f4e143dfd32119c471d0bc43
Reviewed-on: https://chromium-review.googlesource.com/238526
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Trybot-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>