Add a script to reverify DUTs with no recent job history.
The DUT inventory scripts will mark a DUT as "BAD" if it hasn't run
a job recently enough.  This change adds a script that can go through
and find DUTs that could fall into this category, and schedules a
verify job on them, so their status will be known.

BUG=chromium:448241
TEST=Run it with debugging echo commands, to observe it working right.

Change-Id: I6854ccc2119ba4e9a7a8391ddda79cc3ca2502db
Reviewed-on: https://chromium-review.googlesource.com/240311
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>