graphics_SyncControlTest: bring up to date
The commands to start/stop the X server and the way the test result
was being checked were out of date.  The latter was causing failure
to be reported regardless of the actual test result.  Both are changed
to match what other tests currently do, e.g.  graphics_Sanity and
graphics_GLMark.
Also skip this test on freon because the extension tested here
(eglGetSyncValuesCHROMIUM) will not be used under freon.

BUG=chromium:425225
TEST=test_that {lumpy,link_freon} graphics_SyncControlTest

Change-Id: I5af576f15c59b53870ecd55f084be5b1cb31964c
Reviewed-on: https://chromium-review.googlesource.com/242240
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>
Tested-by: Frank Henigman <fjhenigman@chromium.org>