privetd: Add end to end WiFi bootstrapping test
For now, this test merely sets up a router that we'll later attempt to
connect to, and confirms that the DUT is setting up its AP with an SSID
in the format that we expect.  It ends by connecting the router to the
DUTs softAP as a client.

Some work is done to add additional configuration options for
the privetd device model, class, and name.  This allows us
to pick out most of the characters in the SSID.

Some additional work is done to allow the router to connect to
the unencrypted bootstrapping network as a client.

BUG=brillo:8
TEST=This test passes up to the final TestNA indicating that it is
not complete.  network_WiFi_TDLSPing still passes after changes to
add_connected_peer()

Change-Id: Ie46dad4c3442fadf38b7923334d522fb0113037d
Reviewed-on: https://chromium-review.googlesource.com/245462
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>