Autotest: Renamed the control file for all the video related tests
Renaming of the 'Name' field of the contol file will make the
execution of the 'test_that' to get more control over the execution.
Renamed the 'Name' field of the control file to match it to lower case
so as to match the control file naming convention.

Running the specific test would throw error as mentioned in the bug.

BUG=chromium:452733
TEST=Run as 'test_that $hostname video_*.xxx'

Change-Id: Ic3d1bd4b79759af93a90d0e26cd1932b0fa34e6e
Signed-off-by: Ram Mohan Satyanarayana <ram.mohanx.satyanarayana@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/243523
Reviewed-by: Daniel Charles <daniel.charles@intel.com>
Reviewed-by: Sean V Kelley <sean.v.kelley@intel.com>
Reviewed-by: Owen Lin <owenlin@chromium.org>